build: pin the quality gate and fix what it then found
There was no pyproject.toml. Ruff ran against whatever its installed
default rule set happened to be, and neither ruff nor mypy appeared in
requirements.txt. "The lint is clean" therefore meant something
different on every release — a quality gate that moves on its own is not
a gate (NFR-012).

The rule set is now explicit and every ignore carries its reason:

- TRY003 off: this project writes long, specific exception messages on
  purpose. "BANKING at 43.0% exceeds the sector limit of 35.0%" is the
  text a risk manager reads (FR-174); moving it out of the raise to
  please a linter makes the code worse.
- UP042 off: `class X(str, Enum)` is deliberate and documented — members
  must compare equal to the plain strings read back from CHECK-
  constrained SQLite columns. StrEnum changes str() semantics.
- PT018, RUF002/3, E501, PT011 off with reasons in the file.

ruff==0.16.6 and mypy==2.3.1 are pinned alongside the runtime deps.

Turning the gate on properly surfaced two neutered tests, which is the
part worth reading:

- test_fully_supplied_inputs_can_pass ended in
  `assert not r.recomputed.degraded_reason or True`. `X or True` is
  always true, so the line looked like a check and was not one. It now
  asserts what the docstring always claimed — that a complete input set
  reaches a verdict rather than NOT_VALIDATED — and tolerates the
  legitimate degraded flag this fixture raises when its covariance needs
  repairing, while requiring that degradation always explains itself.
- test_post_shock_weights_drift called run_scenario and asserted
  NOTHING. It now checks the shock arithmetic against hand-computed
  values (a 50% fall in half the book is a 25% portfolio loss, 2500
  paise, fully attributed to the asset that fell) and that a scenario
  the control engine could not evaluate is never reported as survived.

Also fixed, all in the same vein:

- zip() without strict= in three places where asset ids are zipped
  against a solver's weight vector. A length mismatch silently truncated
  the portfolio, producing weights that no longer sum to 1 and losing
  assets without a word. Now raises.
- build_constraints computed `n` and never used it; it is now an
  explicit dimension guard, so a mismatch says the vectors disagree
  instead of failing on a broadcast deep inside a cvxpy expression.
- logging.exception passed the exception object it already logs.
- Redundant int(round(...)) casts, unsorted __all__, unused unpackings,
  ambiguous `l`, and non-raw regex in pytest.raises(match=...) where
  "must sum to 1.0" was matching any character for the dot.

mypy is now clean across all 57 source files. 404 tests pass.

Refs: NFR-012, NFR-013, FR-174

+++ cce/portfolio/calculations.py
@@ -101,7 +101,7 @@ def transaction_cost_paise(
                 f"cannot cost a trade in {asset_id}: not in the universe"
             ) from None
         total += rate * abs(delta) * total_value_paise
-    return int(round(total))
+    return round(total)
 
 
 def sector_exposure(
